refactor: extract direct operational fast paths
